에러수정

See merge request s05-webmobile1-sub3/S05P13C103!60

+++ backend/src/main/java/com/ssafy/api/response/dto/Image/ImageResponseDto.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 
 @Getter
 public class ImageResponseDto {
-    private Integer imageId;
+    private Long imageId;
 
     public ImageResponseDto(Image entity){
         this.imageId = entity.getImageId();
